Limity resetowały się ale tylko dla magazynów plantacji, które miały chociaż jedno ukończone zadanie dzienne lub tygodniowe. Zmieniono filtracje dla magazynów plantacji. Resetowanie limitów wydzielono do osobnej metody.
